requirements.txt
requests
Django
djangorestframework
django-cors-headers
django-background-tasks
psycopg2
PyJWT
shortuuid
django_redis
uWSGI
django-crontab
tensorflow
h5py
Keras
pandas
numpy
celery
django-celery-results
gym
